# apps/backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API documentation for IgniticAI backend",
        "title": "IgniticAI API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
"/api/v1/access/check": {
    "get": {
      "summary": "Check user access",
      "description": "Verifies if a user has access to personal or organization assets, with optional admin requirement.",
      "tags": ["access"],
      "produces": ["application/json"],
      "security": [{ "BearerAuth": [] }],
      "parameters": [
        {
          "name": "user_id",
          "in": "path",
          "required": true,
          "type": "string",
          "description": "User ID (UUID)"
        },
        {
          "name": "org_id",
          "in": "query",
          "required": false,
          "type": "string",
          "description": "Organization ID (UUID)"
        },
        {
          "name": "require_admin",
          "in": "query",
          "required": false,
          "type": "boolean",
          "description": "Require admin privileges"
        }
      ],
      "responses": {
        "200": {
          "description": "true if user has access, false otherwise",
          "schema": { "type": "object", "additionalProperties": { "type": "boolean" } }
        },
        "400": { "description": "Invalid request", "schema": { "type": "object" } },
        "401": { "description": "Unauthorized", "schema": { "type": "object" } },
        "403": { "description": "Forbidden", "schema": { "type": "object" } }
      }
    }
  },
  "/api/v1/assets/categories": {
    "get": {
      "summary": "List asset categories",
      "description": "Retrieves a list of all supported asset categories with ID, name, and description.",
      "tags": ["assets"],
      "produces": ["application/json"],
      "security": [{ "BearerAuth": [] }],
      "responses": {
        "200": {
          "description": "List of categories",
          "schema": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": { "type": "string" },
                "name": { "type": "string" },
                "description": { "type": "string" }
              }
            }
          }
        },
        "500": { "description": "Server error", "schema": { "type": "object" } }
      }
    }
  },
  "/api/v1/assets/{id}": {
    "get": {
      "summary": "Get asset details",
      "description": "Retrieves detailed information about a specific asset, including secure URL if stored in Cloudinary.",
      "tags": ["assets"],
      "produces": ["application/json"],
      "security": [{ "BearerAuth": [] }],
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "type": "string",
          "description": "Asset ID (UUID)"
        }
      ],
      "responses": {
        "200": { "description": "Asset details", "schema": { "$ref": "#/definitions/AssetResponse" } },
        "400": { "description": "Invalid asset ID", "schema": { "type": "object" } },
        "401": { "description": "Unauthorized", "schema": { "type": "object" } },
        "403": { "description": "Forbidden", "schema": { "type": "object" } },
        "404": { "description": "Asset not found", "schema": { "type": "object" } },
        "500": { "description": "Server error", "schema": { "type": "object" } }
      }
    },
    "delete": {
      "summary": "Delete an asset",
      "description": "Deletes an asset record and removes its file from Cloudinary if applicable. Only organization admins can delete organization assets.",
      "tags": ["assets"],
      "produces": ["application/json"],
      "security": [{ "BearerAuth": [] }],
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "type": "string",
          "description": "Asset ID (UUID)"
        }
      ],
      "responses": {
        "200": { "description": "Asset deleted successfully", "schema": { "type": "object" } },
        "400": { "description": "Invalid asset ID", "schema": { "type": "object" } },
        "401": { "description": "Unauthorized", "schema": { "type": "object" } },
        "403": { "description": "Forbidden", "schema": { "type": "object" } },
        "404": { "description": "Asset not found", "schema": { "type": "object" } },
        "500": { "description": "Server error", "schema": { "type": "object" } }
      }
    }
  },
  "/api/v1/assets": {
  "get": {
    "summary": "List assets",
    "description": "Lists all assets belonging to a user or organization, with optional category filter.",
    "tags": ["assets"],
    "produces": ["application/json"],
    "security": [{ "BearerAuth": [] }],
    "parameters": [
      {
        "name": "organization_id",
        "in": "query",
        "required": false,
        "type": "string",
        "description": "Organization ID (UUID)"
      },
      {
        "name": "category",
        "in": "query",
        "required": false,
        "type": "string",
        "description": "Asset category filter (business_profile, brand_assets, marketing_assets, analytics_reports, policy_documents, media_documents)"
      }
    ],
    "responses": {
      "200": {
        "description": "List of assets",
        "schema": {
          "type": "array",
          "items": { "$ref": "#/definitions/AssetResponse" }
        }
      },
      "400": { "description": "Invalid request", "schema": { "type": "object" } },
      "401": { "description": "Unauthorized", "schema": { "type": "object" } },
      "403": { "description": "Forbidden", "schema": { "type": "object" } },
      "500": { "description": "Server error", "schema": { "type": "object" } }
    }
  },
  "post": {
    "summary": "Upload an asset",
    "description": "Uploads a file to Cloudinary and creates a corresponding asset record in the database.",
    "tags": ["assets"],
    "consumes": ["multipart/form-data"],
    "produces": ["application/json"],
    "security": [{ "BearerAuth": [] }],
    "parameters": [
      {
        "name": "user_id",
        "in": "header",
        "required": true,
        "type": "string",
        "description": "User ID (from authentication middleware)"
      },
      {
        "name": "file",
        "in": "formData",
        "required": true,
        "type": "file",
        "description": "File to upload"
      },
      {
        "name": "category",
        "in": "formData",
        "required": true,
        "type": "string",
        "description": "Asset category (business_profile, brand_assets, marketing_assets, analytics_reports, policy_documents, media_documents)"
      },
      {
        "name": "title",
        "in": "formData",
        "required": false,
        "type": "string",
        "description": "Asset title (auto-generated from filename if not provided)"
      },
      {
        "name": "organization_id",
        "in": "formData",
        "required": false,
        "type": "string",
        "description": "Organization ID (UUID)"
      },
      {
        "name": "tags",
        "in": "formData",
        "required": false,
        "type": "array",
        "items": { "type": "string" },
        "description": "Tags associated with the asset"
      },
      {
        "name": "metadata",
        "in": "formData",
        "required": false,
        "type": "object",
        "description": "Additional metadata as JSON"
      }
    ],
    "responses": {
      "201": {
        "description": "Asset created successfully",
        "schema": { "$ref": "#/definitions/Asset" }
      },
      "400": { "description": "Invalid request", "schema": { "type": "object" } },
      "401": { "description": "Unauthorized", "schema": { "type": "object" } },
      "403": { "description": "Forbidden", "schema": { "type": "object" } },
      "500": { "description": "Server error", "schema": { "type": "object" } }
    }
  }
}
    }
}